iwa: Keep track of opened response body producers
Short introduction:
When we read response body from a signed web bundle, the reader doesn't
need to parse it. This allows the caller of the reader to read the
response body directly from the signed web bundle file. Access to the
response body on reader side is provided by DataPipeProducer.

Before this CL the instances of the DataPipeProducer are self owned
and get deleted (and close the file) in the callback after the
reading has been completed. This means that the reader doesn't know
if the file is still being read or not.

After this CL the reader will keep track of the DataPipeProducer
instances. This allows the reader to know when the reading is
completed and the signed web bundle file may be removed
(when there are not active producers).

Besides that the web_bundle_file_data_source{.h, .cc} files were
deleted as we switch to the standard mojo::FileDataSource.



Bug: b/287594073
Test: added unittests

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4868845
Commit-Queue: Oleksandr Peletskyi <peletskyi@google.com>
Auto-Submit: Oleksandr Peletskyi <peletskyi@google.com>
Cr-Commit-Position: refs/heads/main@{#1202469}

+++ chrome/browser/web_applications/test/signed_web_bundle_utils.cc
@@ -8,13 +8,12 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/functional/bind.h"
 #include "base/test/test_future.h"
 #include "chrome/browser/web_applications/isolated_web_apps/signed_web_bundle_reader.h"
-#include "components/web_package/mojom/web_bundle_parser.mojom.h"
 #include "testing/gtest/include/gtest/gtest.h"
 
 namespace web_app {
 
 mojo::ScopedDataPipeConsumerHandle ReadResponseBody(
-    uint32_t response_length,
+    uint64_t response_length,
     base::OnceCallback<void(mojo::ScopedDataPipeProducerHandle producer_handle,
                             base::OnceCallback<void(net::Error net_error)>)>
         read_response_body_callback,
@@ -45,7 +44,7 @@ std::string ReadAndFulfillResponseBody(
 }
 
 std::string ReadAndFulfillResponseBody(
-    uint32_t response_length,
+    uint64_t response_length,
     base::OnceCallback<void(mojo::ScopedDataPipeProducerHandle producer_handle,
                             base::OnceCallback<void(net::Error net_error)>)>
         read_response_body_callback) {
